Replace int by size_t and make output with double

+++ nikitov.dmitriy/M1/calculate_square.hpp
@@ -1,8 +1,10 @@
 #ifndef CALCULATE_SQUARE_HPP
 #define CALCULATE_SQUARE_HPP
 
+#include <cstddef>
+
 namespace nikitov
 {
-  double calculateSquare(int radius, int numberOfThreads, int seed, int tries);
+  double calculateSquare(size_t radius, size_t numberOfThreads, size_t seed, size_t tries);
 }
 #endif
